requirements.txt
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
coverage==5.1
cryptography==2.9.2
distro==1.5.0
docker==4.2.1
docker-compose==1.26.0
dockerpty==0.4.1
docopt==0.6.2
Flask==1.1.2
Flask-Caching==1.9.0
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
texttable==1.6.2
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
